pyproject.toml
[project]
name = "suanli-task-manager"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "uvicorn>=0.30.0",
]
